Refactor `SchedulerBuilder`
