+++ src/schedule/core/index.js
@@ -1,6 +1,6 @@
 export * from './schedule.js'
 export * from './scheduler.js'
 export * from './executable.js'
-export * from './systembuilder.js'
+export * from './schedulerbuilder.js'
 export * from './systemconfig.js'
 export * from './runner.js'
